[ci skip]

+++ include/boost/multiprecision/detail/precision.hpp
@@ -27,7 +27,7 @@ inline BOOST_MP_CXX14_CONSTEXPR unsigned current_precision_of_last_chance_imp(co
    // least-significant-bit, ie the number of bits required to represent the
    // the value assuming we will have an exponent to shift things by:
    //
-   return val.is_zero() ? 1 : 1 + digits2_2_10(msb(abs(val)) - lsb(abs(val)) + 1);
+   return static_cast<unsigned>(val.is_zero() ? 1 : 1 + digits2_2_10(msb(abs(val)) - lsb(abs(val)) + 1));
 }
 template <class B, boost::multiprecision::expression_template_option ET>
 inline BOOST_MP_CXX14_CONSTEXPR unsigned current_precision_of_last_chance_imp(const boost::multiprecision::number<B, ET>& val, const std::integral_constant<int, 2>&)
